fixed tabs discover
