+++ src/app/myfeed/_config/config.ts
@@ -2,37 +2,46 @@ export const MENU_ITEMS = [
   {
     label: "Home",
     icon: "HomeIcon",
+    route: "/",
   },
   {
     label: "Search",
     icon: "SearchIcon",
+    route: "/search",
   },
   {
     label: "Library",
     icon: "MenuBookIcon",
+    route: "/library",
   },
   {
     label: "Featured",
     icon: "StarOutlineIcon",
+    route: "/featured",
   },
   {
     label: "Messages",
     icon: "MessagesIcon",
+    route: "/messages",
   },
   {
     label: "Saved",
     icon: "BookmarkBorderOutlinedIcon",
+    route: "/saved",
   },
   {
     label: "My Deals",
     icon: "ShoppingCartOutlinedIcon",
+    route: "/deals",
   },
   {
     label: "Settings",
     icon: "SettingsOutlinedIcon",
+    route: "/settings",
   },
   {
     label: "Profile",
     icon: "AccountCircleOutlinedIcon",
+    route: "/profile",
   },
 ];
