+++ src/app/myfeed/profile/page.tsx
@@ -1,0 +1,7 @@
+import React from "react";
+
+const Profile = () => {
+  return <div>Profile</div>;
+};
+
+export default Profile;
